Boss Battle Code
The objective of this game is to create a limited time frame for the user to encounter a battle with the boss before it reaches the kingdom. The player must level his team, decided in the beginning, to engage with the enemy. Each option chosen will count as a turn for the boss to get closer to the kingdom.